python_ai/requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-jobspy; python_version < "3.13"
requests
beautifulsoup4
lxml
selenium
